+++ src/window.py
@@ -226,7 +226,7 @@ class CurtailWindow(Gtk.ApplicationWindow):
     def check_extension(self, filename):
         file_type = get_file_type(filename)
         if file_type:
-            return file_type in ('png', 'jpg', 'webp')
+            return file_type in ('png', 'jpg', 'webp', 'svg')
         else:
             return False