+++ src/window.py
@@ -309,7 +309,7 @@ class CurtailWindow(Gtk.ApplicationWindow):
                     license_type=Gtk.License.GPL_3_0,
                     website='https://github.com/Huluti/Curtail',
                     issue_url='https://github.com/Huluti/Curtail/issues/new',
-                    version='1.6.0',
+                    version='1.7.0',
                     developers=[
                         'Hugo Posnic https://github.com/Huluti'
                     ],